chore: move zip into separate file
